+++ plugin/blocks/src/components/dropdown/index.js
@@ -1,5 +1,5 @@
-/** External dependencies
- *
+/**
+ * External dependencies
  */
 import classnames from 'classnames';
 
@@ -14,6 +14,8 @@ import { InnerBlocks, useBlockProps, RichText } from '@wordpress/block-editor';
  * Internal dependencies
  */
 import edit from './edit';
+import save from './save';
+
 //import save from './save';
 import metadata from './block.json';
 
@@ -22,18 +24,5 @@ registerBlockType( metadata.name, {
 	 * @see ./edit.js
 	 */
 	edit,
-	save: ({attributes}) => {
-		const blockProps = useBlockProps.save();
-		const { hideChevron } = attributes;
-		return (
-			<details {...blockProps} className={ classnames('dropdown',
-				blockProps.className,
-				{
-					'hide-chevron': hideChevron,
-				 }
-			)}>
-				<InnerBlocks.Content />
-			</details>
-		)
-	}
+	save,
 } );
